Re #327: All billing operations should be done as long

+++ org.aion.avm.core/src/org/aion/avm/core/types/InternalTransaction.java
@@ -15,7 +15,7 @@ public class InternalTransaction extends Transaction {
     }
 
     @Override
-    public int getBasicCost() {
+    public long getBasicCost() {
         return 0;
     }
